project_panel, markdown_preview: Extract open_for_project_path
Move the async open-and-preview logic into MarkdownPreviewView::open_for_project_path,
reducing the project panel handler to validation + one call.
